refactor: rename OpenZeppelin directories
Renaming open-zeppelin to openzeppelin
